+++ src/main/java/com/github/ichenkaihua/controller/UserController.java
@@ -1,13 +1,11 @@
 package com.github.ichenkaihua.controller;
 
 
-import com.github.ichenkaihua.jopo.BaseResponse;
 import com.github.ichenkaihua.jopo.ErrorResponseEntity;
 import com.github.ichenkaihua.model.User;
 import com.github.ichenkaihua.service.UserService;
 import com.github.ichenkaihua.utils.URIUtils;
 import io.swagger.annotations.*;
-import jdk.nashorn.internal.objects.annotations.Optimistic;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,10 +13,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import static org.springframework.http.MediaType.APPLICATION_ATOM_XML_VALUE;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.APPLICATION_XML_VALUE;
 
@@ -38,13 +34,14 @@ public class UserController {
             response = User.class
     )
     @ApiResponses(value = {
-            @ApiResponse(code = 404, message = "指定id的用户不存在",
+            @ApiResponse(code = 400, message = "指定id的用户不存在",
                     response = ErrorResponseEntity.class)
     })
     public ResponseEntity getUserBYId(@ApiParam(value = "用户id",required = true) @PathVariable int id) {
+        System.out.println("userServ:"+userService);
         User user = userService.getUserById(id);
         if (user == null) {
-            return new ErrorResponseEntity(409, "用户创建失败").toResponseEntity();
+            return ErrorResponseEntity.buildToResponseEntity(1000, "用户不存在");
         }
         return ResponseEntity.ok(user);
     }
@@ -52,14 +49,14 @@ public class UserController {
     @RequestMapping(value = "", method = RequestMethod.POST, consumes = {APPLICATION_JSON_VALUE, APPLICATION_XML_VALUE})
     @ApiOperation(value = "添加用户,会忽略id，用于自动创建", code = 201, response = User.class)
     @ApiResponses(value = {
-            @ApiResponse(code = 409, message = "用户已经存在")
+            @ApiResponse(code = 400, message = "用户已经存在",response = ErrorResponseEntity.class)
     })
     public ResponseEntity addUser(@ApiParam(value = "用户信息") @RequestBody User user) {
         User countUser = new User();
         countUser.setName(user.getName());
         //如果存在，返回错误码
         if (userService.isExist(countUser)) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+            return ErrorResponseEntity.buildToResponseEntity(1000, "手机号已经注册");
         }
         userService.addUser(user);
         return ResponseEntity.created(URIUtils.createURI("/users/{id}", user.getId())).body(user);