+++ src/test/java/com/github/ichenkaihua/test/web/Swagger2MarkupTest.java
@@ -1,0 +1,109 @@
+/*
+ *
+ *  Copyright 2015 the original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *
+ */
+package com.github.ichenkaihua.test.web;
+
+import org.junit.Before;
+import org.junit.Ignore;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.restdocs.JUnitRestDocumentation;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.ContextHierarchy;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+import springfox.documentation.staticdocs.SwaggerResultHandler;
+
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@WebAppConfiguration
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextHierarchy({
+        @ContextConfiguration("classpath:spring-config.xml"),
+        @ContextConfiguration("classpath:spring-mvc-config.xml")
+})
+@Ignore
+public class Swagger2MarkupTest {
+
+    private static final Logger LOG = LoggerFactory.getLogger(Swagger2MarkupTest.class);
+
+    @Autowired
+    private WebApplicationContext context;
+
+    private MockMvc mockMvc;
+
+    @Rule
+    public final JUnitRestDocumentation restDocumentationContextProvider =
+            new JUnitRestDocumentation("build/asciidoc");
+
+
+    @Before
+    public void setUp() {
+        this.mockMvc = MockMvcBuilders.webAppContextSetup(this.context)
+                //.apply(documentationConfiguration(this.restDocumentation))
+                .apply(documentationConfiguration(restDocumentationContextProvider))
+                .build();
+    }
+
+   @Test
+    public void addANewPetToTheStore() throws Exception {
+        this.mockMvc.perform(get("/users")
+                .contentType(MediaType.APPLICATION_JSON))
+                .andDo(print())
+                .andDo(document("获取所有用户", preprocessResponse(prettyPrint())))
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    public void createSpringfoxSwaggerJson() throws Exception {
+        //String designFirstSwaggerLocation = Swagger2MarkupTest.class.getResource("/swagger.yaml").getPath();
+
+        String outputDir = System.getProperty("io.springfox.staticdocs.outputDir");
+        System.out.println(outputDir);
+      //  String out = "/home/chenkh/git/ssm-easy-template/build/asciidoc";
+        MvcResult mvcResult = this.mockMvc.perform(get("/v2/api-docs")
+                //api-docs?group=user-api
+                .param("group","user-api")
+                .accept(MediaType.APPLICATION_JSON_UTF8_VALUE))
+                .andDo(print())
+                .andDo(SwaggerResultHandler.outputDirectory(outputDir).build())
+                .andExpect(status().isOk())
+                .andReturn();
+
+        String springfoxSwaggerJson = mvcResult.getResponse().getContentAsString();
+        System.out.println(springfoxSwaggerJson);
+
+        //SwaggerAssertions.assertThat(Swagger20Parser.parse(springfoxSwaggerJson)).isEqualTo(designFirstSwaggerLocation);
+    }
+
+}